+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,27 @@
+## チケットへのリンク
+
+- https://example.com
+
+## やったこと
+
+- このプルリクで何をしたのか？
+
+## やらないこと
+
+- このプルリクでやらないことは何か？（あれば。無いなら「無し」で OK）（やらない場合は、いつやるのかを明記する。）
+
+## 動作確認
+
+- どのような動作確認を行ったのか？　結果はどうか？
+
+## 動作確認用 URL
+
+- http://localhost:3000/example (あれば。)
+
+## スクリーンショット
+
+- UI の変更があれば実装前後の比較スクリーンショットを添付
+
+## その他
+
+- レビュワーへの参考情報（実装上の懸念点や注意点などあれば記載）
