+++ frontend/src/components/features/reception/presenter/Reception.tsx
@@ -4,32 +4,14 @@ import {
   Card,
   CardBody,
   Checkbox,
-  Drawer,
-  DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerOverlay,
   HStack,
   Heading,
-  IconButton,
   Image,
-  Input,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
   Stack,
   VStack,
-  useDisclosure,
 } from '@chakra-ui/react';
-import { Ticket } from '../../../ui/icon';
-import { useRef } from 'react';
-import React from 'react';
+import OrderConfirmationModal from './OrderConfirmationModal';
+import TicketIdInputDrawer from './TicketIdInputDrawer';
 
 const ReceptionUI = () => {
   return (
@@ -55,12 +37,12 @@ const ReceptionUI = () => {
             <Stack>
               <CardBody>
                 <Heading size="lg">ソース＆マヨ</Heading>
-                <VStack align="stretch">
+                {/* <VStack align="stretch">
                   <Checkbox defaultChecked>ソース</Checkbox>
                   <Checkbox defaultChecked>マヨネーズ</Checkbox>
                   <Checkbox defaultChecked>かつお節</Checkbox>
                   <Checkbox defaultChecked>青のり</Checkbox>
-                </VStack>
+                </VStack> */}
                 <Button>カートに追加</Button>
               </CardBody>
             </Stack>
@@ -143,121 +125,11 @@ const ReceptionUI = () => {
         </Card>
       </HStack>
       <HStack justify="center" mb="10">
-        <ReturnFocus />
+        <OrderConfirmationModal />
       </HStack>
-      <DrawerExample />
+      <TicketIdInputDrawer />
     </Box>
   );
 };
-
-function DrawerExample() {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const btnRef = useRef();
-
-  return (
-    <>
-      <IconButton
-        aria-label="ticket"
-        position="fixed"
-        bottom="40px"
-        right="40px"
-        icon={<Ticket />}
-        opacity={0.7}
-        isRound={true}
-        fontSize="300px"
-        size={{ base: '100px', sm: '150px' }}
-        ref={btnRef.current}
-        onClick={onOpen}
-      />
-      <Drawer
-        size="lg"
-        isOpen={isOpen}
-        placement="right"
-        onClose={onClose}
-        finalFocusRef={btnRef.current}
-      >
-        <DrawerOverlay />
-        <DrawerContent>
-          <DrawerCloseButton />
-          <DrawerHeader>前売り券番号を入力してください</DrawerHeader>
-
-          <DrawerBody>
-            <Input placeholder="入力して..." />
-          </DrawerBody>
-
-          <DrawerFooter>
-            <Button variant="outline" mr={3} onClick={onClose}>
-              閉じる
-            </Button>
-            <Button colorScheme="blue">検索</Button>
-          </DrawerFooter>
-        </DrawerContent>
-      </Drawer>
-    </>
-  );
-}
-
-function ReturnFocus() {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const finalRef = React.useRef(null);
-
-  return (
-    <>
-      <Button size="lg" onClick={onOpen}>
-        注文を確定
-      </Button>
-      <Modal finalFocusRef={finalRef} isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>注文内容</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <VStack justify="center">
-              <Card
-                direction={{ base: 'column', sm: 'row' }}
-                overflow="hidden"
-                variant="outline"
-                width={{ base: '100%', sm: 'full' }}
-                height={{ base: '100%', sm: '100px' }}
-                borderRadius={30}
-              ></Card>
-              <Card
-                direction={{ base: 'column', sm: 'row' }}
-                overflow="hidden"
-                variant="outline"
-                width={{ base: '100%', sm: 'full' }}
-                height={{ base: '100%', sm: '100px' }}
-                borderRadius={30}
-              ></Card>
-              <Card
-                direction={{ base: 'column', sm: 'row' }}
-                overflow="hidden"
-                variant="outline"
-                width={{ base: '100%', sm: 'full' }}
-                height={{ base: '100%', sm: '100px' }}
-                borderRadius={30}
-              ></Card>
-              <Card
-                direction={{ base: 'column', sm: 'row' }}
-                overflow="hidden"
-                variant="outline"
-                width={{ base: '100%', sm: 'full' }}
-                height={{ base: '100%', sm: '100px' }}
-                borderRadius={30}
-              ></Card>
-            </VStack>
-          </ModalBody>
-
-          <ModalFooter>
-            <Button colorScheme="blue" mr={3} onClick={onClose}>
-              注文をせずに戻る
-            </Button>
-            <Button variant="ghost">注文を確定する</Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
-    </>
-  );
-}
 
 export default ReceptionUI;
